Add module structure to use pytest
